backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["santiago <smelomedina05@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6"
Flask = "1.1.2"
mysql-connector-python = "8.0.23"
pyjwt = "2.0.1"

[tool.poetry.dev-dependencies]
Flask-Cors = "3.0.9"
Flask = "1.1.2"
mysql-connector-python = "8.0.23"
pyjwt = "2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
